refactor: group shared utilities into lib/utils/
