+++ src/app/api/images/route.ts
@@ -59,7 +59,7 @@
 
 import { NextRequest, NextResponse } from 'next/server';
 import { prisma } from '@/lib/prisma';
-import { serializeForJson } from '@/lib/type-utils';
+import { serializeForJson } from '@/lib/utils/type-utils';
 import { apiError } from '@/lib/api-utils';
 
 export async function GET(request: NextRequest) {
