+++ src/app/api/scans/[id]/scanner-results/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server'
 import { prisma } from '@/lib/prisma'
-import { serializeForJson } from '@/lib/type-utils'
+import { serializeForJson } from '@/lib/utils/type-utils'
 import { apiError } from '@/lib/api-utils'
 
 export async function GET(
