+++ src/app/api/scans/aggregated/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server'
 import { prisma } from '@/lib/prisma'
-import { getSeverityWeight } from '@/lib/severity-utils'
+import { getSeverityWeight } from '@/lib/utils/severity-utils'
 import { apiError } from '@/lib/api-utils'
 
 export async function GET(request: NextRequest) {
