+++ src/components/scan/VulnerabilityDetailModal.tsx
@@ -23,9 +23,9 @@ import {
   IconFingerprint,
   IconLink,
 } from "@tabler/icons-react";
-import { formatDate, renderValue } from "@/lib/format-utils";
+import { formatDate, renderValue } from "@/lib/utils/format-utils";
 import { copyToClipboard } from "@/lib/clipboard";
-import { getSeverityCssClass } from "@/lib/severity-utils";
+import { getSeverityCssClass } from "@/lib/utils/severity-utils";
 
 interface VulnerabilityDetailModalProps {
   isOpen: boolean;
