+++ src/components/scan/findings/VulnerabilitiesTab.tsx
@@ -27,7 +27,7 @@ import {
   TableRow,
 } from "@/components/ui/table";
 import { VulnerabilityUrlMenu } from "@/components/vulnerability-url-menu";
-import { getSeverityCssClass } from "@/lib/severity-utils";
+import { getSeverityCssClass } from "@/lib/utils/severity-utils";
 
 interface VulnerabilitiesTabProps {
   vulnerabilities: any[];
