+++ src/components/vulnerability-details-modal.tsx
@@ -26,9 +26,9 @@ import {
   IconCheck,
   IconCopy,
 } from "@tabler/icons-react";
-import { getSeverityBadgeVariant } from "@/lib/severity-utils";
+import { getSeverityBadgeVariant } from "@/lib/utils/severity-utils";
 import { copyToClipboard } from "@/lib/clipboard";
-import { getImageName } from "@/lib/image-utils";
+import { getImageName } from "@/lib/utils/image-utils";
 
 interface VulnerabilityDetailsModalProps {
   vulnerability: {
