+++ src/components/vulnerability-selection-modal.tsx
@@ -16,7 +16,7 @@ import { ScrollArea } from '@/components/ui/scroll-area';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { Loader2 } from 'lucide-react';
-import { getSeverityCssClass } from '@/lib/severity-utils';
+import { getSeverityCssClass } from '@/lib/utils/severity-utils';
 
 interface Vulnerability {
   id: string;
