+++ src/hooks/useScans.ts
@@ -1,6 +1,6 @@
 import { useApp } from '@/contexts/AppContext'
 import { useMemo } from 'react'
-import { getImageName } from '@/lib/image-utils'
+import { getImageName } from '@/lib/utils/image-utils'
 
 export function useScans() {
   const { state, dispatch } = useApp()
