+++ src/lib/utils/image-utils.ts
