+++ src/lib/utils/severity-utils.ts
